python/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

setuptools==65.5.1
pip>=21
apsw<3.10
pkginfo==1.7.1
beautifultable==1.0.0
cachetools==3.0.0
casbin==1.16.6
casbin-sqlalchemy-adapter==0.4.2
cloudpickle==2.1.0
cos-python-sdk-v5==1.9.10
Flask==2.0.3
gmpy2==2.0.8
joblib==1.2.0
kazoo==2.6.1
pyyaml==5.4.1
kfserving==0.5.1
kubernetes>=12.0.1
lmdb==1.3.0
minio>=6.0.2, <7.0.0
numpy==1.23.1
peewee==3.9.3
psutil>=5.7.0
pycryptodomex==3.15.0
PyMySQL==1.0.3
python-dotenv==0.13.0
redis==3.5.3
urllib3==1.26.5
requests<2.26.0
tensorflow-estimator==2.9.0rc0
requests_toolbelt==0.9.1
ruamel-yaml==0.16.10
scikit-learn==1.0.1
scipy==1.8.1
tensorflow-cpu==2.9.3
#torch==1.13.1+cpu
torchvision==0.14.1+cpu
torch-model-archiver==0.3.1
transformers==4.23.1
coverage==4.5.2
pika==1.2.1
pandas==1.1.5
cryptography==3.3.2
sortedcontainers==2.2.2
pytorch-lightning>=1.6.5
filelock==3.3.1
pulsar-client==2.10.2
fastavro==1.4.1
lightgbm==3.3.1
shortuuid==1.0.9
sklearn2pmml==0.90.0
nyoka==5.3.0
transformers==4.23.1
impyla==0.18.0
venv-pack==0.2.0
fate-crypto==0.3.0
ipcl-python==2.0.0
mldb

# addtional requirements for eggroll
grpcio==1.48.0
grpcio-tools==1.48.0
numba==0.53.0
protobuf==3.19.6
pyarrow==6.0.1
mmh3==3.0.0

# addtional requirements for fate-client
click>=7.1.2,<8.0.0
colorama>=0.4.4
loguru>=0.6.0
poetry>=0.12
prettytable>=1.0.0,<2.0.0
#setuptools>=50.0,<51.0
sshtunnel>=0.1.5,<0.2.0
packaging<21.0,>=20.4


# web_fl
itsdangerous==2.0.1
hdfs
Flask-Login
Flask-SocketIO
paramiko
flask-cors
cx_Oracle
impala
impyla
happybase
apscheduler
psycopg2-binary

Werkzeug==2.0.0
pyspark==3.3.2
func_timeout

fastparquet==0.7.1

snowland-smx==0.3.1
gmssl==3.2.2
pysmx==0.4.0
cache3==0.2.1
cython
backports.lzma
torch==2.0
